+++ SharedModel/Entity/SimpleDTO.cs
@@ -4,11 +4,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace SharedModel.ServiceStackFolderModel
+namespace SharedModel.Entity
 {
     public class SimpleDTO
     {
         public long Id { get; set; }
-        public string MyProperty { get; set; }
+        public string Name { get; set; }
     }
 }
