+++ SharedModel/ServiceInterface/GetSimpleDTO.cs
@@ -1,13 +1,15 @@
 ﻿using ServiceStack;
+using SharedModel.Entity;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace SharedModel.ServiceStackFolderModel
+namespace SharedModel.ServiceInterface
 {
     [Route("/ServiceStack")]
     [Route("/ServiceStack/{Name}")]
     public class GetSimpleDTO : IReturn<List<SimpleDTO>>
     {
+        public string Name { get; set; }
     }
 }
